网页改为了读配置文件。玩了一把模板。删掉了cximage和jsoncpp，因为相关依赖已经被我丢到了dll里。我还需要把svr两个垃圾干掉。

+++ examples/MFCApplication1/FileSystem.cpp
@@ -39,7 +39,7 @@ namespace Alime
 			{
 				std::vector<wchar_t> buffer(fullPath.length() + 1);
 				wcscpy_s(&buffer[0], fullPath.length() + 1, fullPath.c_str());
-				for (int i = 0; i < buffer.size(); i++)
+				for (size_t i = 0; i < buffer.size(); i++)
 				{
 					if (buffer[i] == L'\\' || buffer[i] == L'/')
 					{
